6.5_full_text_search.rb created, test_milestones sort_by to_f,  milestone 6_list_items pass

+++ test/test_milestones.rb
@@ -15,7 +15,7 @@ class Milestones < Test::Unit::TestCase
   def test_milestones
     milestone_path = $own_path + "/milestones"
     passed, failed = Dir[milestone_path + "/*.rb"].
-                      sort_by{|ms| File.basename(ms).to_i }.
+                      sort_by{|ms| File.basename(ms).to_f }.
                       partition{|ms| check_milestone(ms) }
 
     puts
